move typedefs and resolvers to own file
